backend/requirements.txt
fastapi
uvicorn[standard]
openai
arxiv
httpx
pydantic
python-dotenv
